refactored + removed unnecessary files
